requirements.txt
opencv-python~=4.7.0.72
numpy~=1.21.5
torchvision~=0.14.0
matplotlib~=3.5.1
simpleitk~=2.2.1
tqdm~=4.64.0
pillow~=9.0.1
scikit-image~=0.19.2
monai~=1.3.0
medpy~=0.4.0
scipy~=1.7.3
scikit-learn~=1.0.2